dedicated test file for tabular-fits
